+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # `SwipProteomics/CHANGELOG.md`
 
+#### Thu Sep 17 10:56:29 EDT 2020
+* install NBGOF package - need to evaluate GOF of NB model
+* added PSM report to data
+
 #### Tue Sep 15 09:33:40 EDT 2020
 * generate sample pca showing good resolution of different subcellular fractions
 * mkdir `docs/review` to init work on reviews